oss-fuzz: Fix undefined-behavior
Analyzer run time does not like A -= B or A +=B even if A and B are uint32_t as there is potential
for overflow.

Analyzer run time does not like overflow when doing coap_tick_t - coap_tick_t.

Analyzer run time does not like (uint8_t << 24) as uint8_t is promoted as int, not u_int.

Analyzer run time does not like while(A--) {} even if A is uint32_t, as A is always decremented on
the test, even if the value of A is 0.

Fix potential overflow when analyzing an asn.1 tag.

Fix potential overflow when doing decode_segment() on a % hex value.

+++ tests/oss-fuzz/network_message_target.c
@@ -91,14 +91,14 @@ test_context_config(coap_context_t *ctx, const uint8_t *data, size_t size) {
   coap_context_set_max_token_size(ctx, token_size);
 
   if (size >= 20) {
-    size_t csm_max_message_size = (data[16] << 24) | (data[17] << 16) |
+    size_t csm_max_message_size = ((uint32_t)data[16] << 24) | (data[17] << 16) |
                                   (data[18] << 8) | data[19];
     /* CSM max message size must be >= 64 per spec */
     if (csm_max_message_size >= 64) {
       coap_context_set_csm_max_message_size(ctx, csm_max_message_size);
     }
 
-    uint32_t timeout = (data[12] << 24) | (data[13] << 16) |
+    uint32_t timeout = ((uint32_t)data[12] << 24) | (data[13] << 16) |
                        (data[14] << 8) | data[15];
     coap_context_set_csm_timeout_ms(ctx, timeout);
   }
